gitflow-feature-stash: fix-deadlock

+++ db/cockroachdb/seeder/post_seeder.go
@@ -1,8 +1,10 @@
 package seeder
 
 import (
+	"context"
 	"social-media/internal/config"
 	"social-media/test/mock"
+	"time"
 )
 
 type PostSeeder struct {
@@ -22,13 +24,20 @@ func NewPostSeeder(
 }
 
 func (postSeeder *PostSeeder) Up() {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
 	for _, post := range postSeeder.PostMock.Data {
-		begin, beginErr := postSeeder.DatabaseConfig.CockroachdbDatabase.Connection.Begin()
+		connection, acquireErr := postSeeder.DatabaseConfig.CockroachDatabase.Pool.Acquire(ctx)
+		if acquireErr != nil {
+			panic(acquireErr)
+		}
+		begin, beginErr := connection.Begin(ctx)
 		if beginErr != nil {
 			panic(beginErr)
 		}
 
 		_, err := begin.Query(
+			ctx,
 			"INSERT INTO \"post\" (id, user_id, image_url, description, created_at, updated_at, deleted_at) VALUES ($1, $2, $3, $4, $5, $6, $7);",
 			post.Id,
 			post.UserId,
@@ -42,31 +51,42 @@ func (postSeeder *PostSeeder) Up() {
 			panic(err)
 		}
 
-		commitErr := begin.Commit()
+		commitErr := begin.Commit(ctx)
 		if commitErr != nil {
 			panic(commitErr)
 		}
+
+		connection.Release()
 	}
 }
 
 func (postSeeder *PostSeeder) Down() {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
 	for _, post := range postSeeder.PostMock.Data {
-		begin, beginErr := postSeeder.DatabaseConfig.CockroachdbDatabase.Connection.Begin()
+		connection, acquireErr := postSeeder.DatabaseConfig.CockroachDatabase.Pool.Acquire(ctx)
+		if acquireErr != nil {
+			panic(acquireErr)
+		}
+		begin, beginErr := connection.Begin(ctx)
 		if beginErr != nil {
 			panic(beginErr)
 		}
 
-		_, err := begin.Query(
+		_, queryErr := begin.Query(
+			ctx,
 			"DELETE FROM \"post\" WHERE id = $1 LIMIT 1;",
 			post.Id,
 		)
-		if err != nil {
-			panic(err)
+		if queryErr != nil {
+			panic(queryErr)
 		}
 
-		commitErr := begin.Commit()
+		commitErr := begin.Commit(ctx)
 		if commitErr != nil {
 			panic(commitErr)
 		}
+
+		connection.Release()
 	}
 }
